+++ algebra/algebralib/src/ErrorHandling.cpp
@@ -31,7 +31,7 @@ namespace Algebra{
 	/********************************************************/
 
 #ifdef DEBUG
-	const static size_t MAX_FMT = 256;
+	const static uint64_t MAX_FMT = 256;
 	::std::string ALGEBRALIB_FMT(const char* format, ...) {
 		char buf[MAX_FMT];
 		va_list args;
